[file_history] Fix error of how to read git sha
Using the fs and traverse through the "git log -p -- filename" removes
all informations after a renamed of the file. So we loose a lot of data.

The new way is to use pydriller modifiedFiles that take into account renames, add, deletion etc...
From this, can get info of the modifiy files through all the commits of the repo.

Note: Rename of a file, is not that the file name has changed, but the path of the file has changed (moved folder)

More info:
https://pydriller.readthedocs.io/en/latest/modifiedfile.html